requirements.txt
# Core Dependencies
flask
flask-cors
flask-sqlalchemy
flask-migrate
flask-login
flask-wtf
flask-bcrypt
flask-mail
sqlalchemy
faker
python-dotenv
schedule
requests
datetime
uuid
bcrypt
langchain
langchain-groq
langchain-community
langgraph
langchain-core
langchain-huggingface
chromadb
langchain-faiss
langchain_openai
langchain-experimental
langgraph
sentence-transformers
asyncio
aiohttp
asyncpg
pymysql
pandas
numpy
matplotlib
seaborn
plotly
scikit-learn
scipy
